railway-requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
apscheduler==3.10.4
pandas==2.0.3
requests==2.31.0
web3==6.11.0
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1